refactor: split admin database section
